+++ spdcalc/src/math/nelder_mead.rs
@@ -36,7 +36,7 @@ use optimize::*;
 pub fn nelder_mead_1d(
   func : impl Fn(f64) -> f64,
   guess : f64,
-  max_iter : u32,
+  max_iter : usize,
   min : f64,
   max : f64,
   tolerance : f64,
